Add support for null values on some built-in equality checks (#15)
Extends a few of the built-in equality checks to also accept `null` and `undefined` as values. It is not uncommon for selectors or maps to have "optional" return values. So having it supported in the built-in checks makes them available to be used in more situations.

Added new the checks:
- `nullableShallowObjectEqualityCheck`
- `nullableShallowObjectArrayEqualityCheck`
- `nullableShallowArrayEqualityCheck`