Add support for multiple facets in useFacetCallback and useFacetEffect (#14)
This was been motivated by internal use-cases that relied on adding extra `useFacetMaps` to achieve the same goal. This should have a lower overhead with the added benefit that now all of our hooks share the same API (facets are arrays).

Also updates the `useRemoteFacetPropSetter` to accept a `remoteFacet` instead of a facet, making the API more consistent. I did this update as I needed to touch the implementation.