Alternative proposal to support shared facet error
Disclaimer: this is not a complete implementation, but rather a piece for discussion.

## Native facet error support

Given the `createFacet` is already aware of the subscriptions, it should support natively the error handling. So, this introduces a `onError` method to facets.

This new `onError` currently is only triggered when subscriptions fail, but I could see it being extended in the future for more error handling.

The main benefit of moving it as facet method, is that then handling the error while subscribing is much simpler.

So, within the `useSharedFacet` it can just setup the new error callback within an effect.

## ErrorBoundary

I also believe we should rely on native error handling of React for this as a fallback for scenarios that is not handled explicitly when using `useSharedFacet`'s second `onError` callback.

My main motivation is that if there is an error such as this, we will probably already need to unmount areas of the screen, so reconciliation is probably unavoidable. So better to rely on existing React infrastructure.

## Memoization of shared facets

This is a known limitation of the current "memoization" solution that it never cleans-up. However, given "shared facets" are static, and their definitions never grow within an App, its not something of a current concern.

Moving towards a future where we would accept "parameterized shared facets", then we will need a more robust solution. But once that arrives, it should probably be handled by relying on React's lifecycle.

I've done some experiment within these lines that can be found in the [experimenting-with-shared-facets branch](https://github.com/Mojang/ore-ui/compare/experimenting-with-shared-facets).